Charts & config move
