+++ frontend/app/components/samples/heat-map-chart.sample.vue
@@ -8,7 +8,7 @@
 import LfxChart from '~/components/uikit/chart/Chart.vue';
 import type { ChartData, ChartSeries, CategoryData } from '~/components/uikit/chart/types/ChartTypes';
 import { getHeatMapChartConfig } from '~/components/uikit/chart/configs/heat-map.chart';
-import { lfxColors } from '~/components/config/styles/colors';
+import { lfxColors } from '~/config/styles/colors';
 
 const props = defineProps<{
   chartData: ChartData[];
