+++ frontend/app/components/samples/scatter-chart.sample.vue
@@ -8,7 +8,7 @@
 import LfxChart from '~/components/uikit/chart/Chart.vue';
 import type { ChartData, ChartSeries, CategoryData } from '~/components/uikit/chart/types/ChartTypes';
 import { getScatterChartConfig } from '~/components/uikit/chart/configs/scatter.chart';
-import {lfxColors} from "~/components/config/styles/colors";
+import {lfxColors} from "~/config/styles/colors";
 
 const props = defineProps<{
   chartData: ChartData[];
