+++ frontend/app/components/shared/layout/search.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    class="w-full max-w-80 overflow-hidden bg-neutral-50 h-8 border
+    class="w-full md:max-w-80 overflow-hidden bg-neutral-50 h-8 border
   border-neutral-200 rounded-full flex items-center gap-2 px-4">
     <lfx-icon name="search" class="text-neutral-400" :size="14" />
     <p class="text-body-1 text-neutral-400 truncate">
